Centralize notificationDelegate and add pixel firing
