+++ iOS/DuckDuckGo/AppServices/VPNService.swift
@@ -30,6 +30,7 @@ final class VPNService: NSObject {
     private let tunnelDefaults = UserDefaults.networkProtectionGroupDefaults
     private let vpnFeatureVisibility: DefaultNetworkProtectionVisibility = AppDependencyProvider.shared.vpnFeatureVisibility
     private let tipKitAppEventsHandler = TipKitAppEventHandler()
+    private let notificationServiceManager: NotificationServiceManager
 
     private let mainCoordinator: MainCoordinator
     private let subscriptionManager: any SubscriptionAuthV1toV2Bridge
@@ -41,9 +42,10 @@ final class VPNService: NSObject {
         self.mainCoordinator = mainCoordinator
         self.subscriptionManager = subscriptionManager
         self.application = application
+        self.notificationServiceManager = NotificationServiceManager(mainCoordinator: mainCoordinator)
+        notificationCenter.delegate = notificationServiceManager
+        
         super.init()
-
-        notificationCenter.delegate = self
 
         widgetRefreshModel.beginObservingVPNStatus()
         tipKitAppEventsHandler.appDidFinishLaunching()
@@ -128,30 +130,6 @@ final class VPNService: NSObject {
                                       icon: UIApplicationShortcutIcon(templateImageName: "VPN-16"),
                                       userInfo: nil)
         ]
-    }
-
-}
-
-extension VPNService: UNUserNotificationCenterDelegate {
-
-    func userNotificationCenter(_ center: UNUserNotificationCenter,
-                                willPresent notification: UNNotification,
-                                withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
-        completionHandler(.banner)
-    }
-
-    @MainActor
-    func userNotificationCenter(_ center: UNUserNotificationCenter,
-                                didReceive response: UNNotificationResponse,
-                                withCompletionHandler completionHandler: @escaping () -> Void) {
-        if response.actionIdentifier == UNNotificationDefaultActionIdentifier {
-            let identifier = response.notification.request.identifier
-
-            if NetworkProtectionNotificationIdentifier(rawValue: identifier) != nil {
-                mainCoordinator.presentNetworkProtectionStatusSettingsModal()
-            }
-        }
-        completionHandler()
     }
 
 }
