+++ iOS/DuckDuckGo/NotificationServiceManager.swift
@@ -1,0 +1,58 @@
+//
+//  NotificationServiceManager.swift
+//  DuckDuckGo
+//
+//  Copyright © 2025 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import VPN
+import Subscription
+import UIKit
+import NotificationCenter
+import Core
+
+final class NotificationServiceManager: NSObject, UNUserNotificationCenterDelegate {
+    
+    private let mainCoordinator: MainCoordinator
+    
+    init(mainCoordinator: MainCoordinator) {
+        self.mainCoordinator = mainCoordinator
+    }
+    
+    func userNotificationCenter(_ center: UNUserNotificationCenter,
+                                willPresent notification: UNNotification,
+                                withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
+        completionHandler(.banner)
+    }
+
+    @MainActor
+    func userNotificationCenter(_ center: UNUserNotificationCenter,
+                                didReceive response: UNNotificationResponse,
+                                withCompletionHandler completionHandler: @escaping () -> Void) {
+        if response.actionIdentifier == UNNotificationDefaultActionIdentifier {
+            let identifier = response.notification.request.identifier
+
+            if NetworkProtectionNotificationIdentifier(rawValue: identifier) != nil {
+                mainCoordinator.presentNetworkProtectionStatusSettingsModal()
+            }
+            
+            if identifier == InactivityNotificationSchedulerService.notificationIdentifier {
+                Pixel.fire(pixel: .provisionalPushNotificationTapped)
+            }
+        }
+        completionHandler()
+    }
+    
+}
